The rest of the signed shifts, in six files
The same finding as display.hxx, everywhere else it occurs. Fifteen sites, and
every one of them is a narrow unsigned type promoting to a SIGNED int before
the shift - which is the part that is invisible in the source:

    static_cast<UInt16>(f[7]) << 8      UInt16 promotes to int
    (bits >> row)                       UInt8 promotes to int, row is Int32
    (a >> 11) & 0x1F                    a is UInt16, so this is int arithmetic
    ~st->lamps                          UInt8 promotes, then ~ makes it negative

Nothing here misbehaves today. Every shift count is a small constant, every
value fits, and where a value could be negative the caller has already clamped
it. What changes is that the conversion is now written down instead of being a
rule you have to remember.

  cue.hxx             the lamp mask, where ~ on a promoted UInt8 gives a
                      negative int that only survives because it is cast back
  dfplayer.hxx        three big-endian pairs: the folder/track argument and
                      the two reply fields
  dfplayer_proto.hxx  the frame's parameter and checksum high bytes
  storage.hxx         CSD v1 mult, which was already casting the OUTSIDE of
                      the expression while the arithmetic inside stayed signed
  gfx.hxx             blend() unpacking RGB565, and the glyph bit test
  hal.hxx             readReg16U16 assembling its two bytes

blend() converts once into au/bu rather than casting six times, which is the
same reason putCoord exists.

Repo-wide: 15 real to 0. The control is that the check still reports 32 on
these files - the literal `8` in `>> 8` is an int and always will be - so a
zero here is a zero from a check that ran.

verify PASS: both boards 0 errors 0 warnings 3 images, seven suites, audit 0.

+++ firmware/lib/drivers/dfplayer_proto.hxx
@@ -150,7 +150,8 @@ namespace bibo::dfplayer
    */
   inline Void frame(UInt8* out, const UInt8 cmd, const UInt8 ack, const UInt16 param)
   {
-  const auto paramHi = static_cast<UInt8>((param >> 8) & 0xFFu);
+  const auto paramHi = static_cast<UInt8>((static_cast<UInt32>(param) >> 8u)
+                                          & 0xFFu);
   const auto paramLo = static_cast<UInt8>(param & 0xFFu);
 
   const auto sum = static_cast<UInt16>(0xFFu + 0x06u + cmd + ack
@@ -164,7 +165,7 @@ namespace bibo::dfplayer
   out[4] = ack;
   out[5] = paramHi;
   out[6] = paramLo;
-  out[7] = static_cast<UInt8>((chk >> 8) & 0xFFu);
+  out[7] = static_cast<UInt8>((static_cast<UInt32>(chk) >> 8u) & 0xFFu);
   out[8] = static_cast<UInt8>(chk & 0xFFu);
   out[9] = 0xEFu;
   }
